fix: исправить ошибки тест 1

+++ controllers/cards.js
@@ -26,41 +26,53 @@ const getCards = (req, res) => {
 };
 
 const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .then((card) => {
-      if (!card) {
-        res.status(404).send({ message: 'Карточка не найдена.' });
-        return;
-      }
-      res.send({ message: 'Карточка удалена.' });
-    })
-    .catch(() => res.status(404).send({ message: 'Карточка не найдена.' }));
+  if (req.params.cardId.length === 24) {
+    Card.findByIdAndRemove(req.params.cardId)
+      .then((card) => {
+        if (!card) {
+          res.status(404).send({ message: 'Карточка не найдена.' });
+          return;
+        }
+        res.send({ message: 'Карточка удалена.' });
+      })
+      .catch(() => res.status(404).send({ message: 'Карточка не найдена.' }));
+  } else {
+    res.status(400).send({ message: 'Некорректный id карточки.' });
+  }
 };
 
 const likedCard = (req, res) => {
-  Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
-    .populate(['owner', 'likes'])
-    .then((card) => {
-      if (!card) {
-        res.status(404).send({ message: 'Карточка не найдена.' });
-        return;
-      }
-      res.send(card);
-    })
-    .catch(() => res.status(404).send({ message: 'Карточка не найдена.' }));
+  if (req.params.cardId.length === 24) {
+    Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
+      .populate(['owner', 'likes'])
+      .then((card) => {
+        if (!card) {
+          res.status(404).send({ message: 'Карточка не найдена.' });
+          return;
+        }
+        res.send(card);
+      })
+      .catch(() => res.status(404).send({ message: 'Карточка не найдена.' }));
+  } else {
+    res.status(400).send({ message: 'Некорректный id карточки.' });
+  }
 };
 
 const dislikedCard = (req, res) => {
-  Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
-    .populate(['owner', 'likes'])
-    .then((card) => {
-      if (!card) {
-        res.status(404).send({ message: 'Карточка с указанным _id не найдена.' });
-        return;
-      }
-      res.send(card);
-    })
-    .catch(() => res.status(404).send({ message: 'Карточка с указанным _id не найдена.' }));
+  if (req.params.cardId.length === 24) {
+    Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
+      .populate(['owner', 'likes'])
+      .then((card) => {
+        if (!card) {
+          res.status(404).send({ message: 'Карточка с указанным _id не найдена.' });
+          return;
+        }
+        res.send(card);
+      })
+      .catch(() => res.status(404).send({ message: 'Карточка с указанным _id не найдена.' }));
+  } else {
+    res.status(400).send({ message: 'Некорректный id карточки.' });
+  }
 };
 
 module.exports = {
